requirements.txt
numpy==2.3.4
pandas==2.3.3
matplotlib==3.10.7
scikit-learn==1.7.2
scipy==1.16.3
pygame==2.6.1
python-dotenv==1.0.0

